twitter_inec_project/requirements.txt
prefect
pandas
tweepy
gcloud
dbt-core
dbt-bigquery
python-dotenv

